requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
blinker==1.9.0
build==1.3.0
buildozer==1.5.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Cython==0.29.33
distlib==0.4.0
filelock==3.18.0
Flask==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==5.29.5
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
requests==2.32.4
rsa==4.9.1
sh==1.14.3
SpeechRecognition==3.14.3
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.33.1
Werkzeug==3.1.3
